fix: Editor fixes
