+++ packages/editor/src/components/blocks/FileBlock.tsx
@@ -13,6 +13,7 @@ import {
   File as FileIcon
 } from 'lucide-react'
 import { useState } from 'react'
+import { saveFileToPublic } from '@/lib/fileUtils'
 
 interface FileBlockProps {
   content: string
@@ -60,6 +61,27 @@ const formatFileType = (fileType: string) => {
 export function FileBlock({ content, id, onUpdate, isEditing, metadata }: FileBlockProps) {
   const [isHovered, setIsHovered] = useState(false)
 
+  const handleFileUpload = async (e: React.ChangeEvent<HTMLInputElement>) => {
+    e.stopPropagation()
+    const file = e.target.files?.[0]
+    if (!file) return
+
+    try {
+      const filename = await saveFileToPublic(file, content)
+      
+      const fileContent = JSON.stringify({
+        url: `/${filename}`,
+        name: file.name,
+        fileType: file.type,
+        size: file.size
+      })
+
+      onUpdate?.(fileContent)
+    } catch (error) {
+      console.error('Failed to upload file:', error)
+    }
+  }
+
   const parseFileContent = (content: string) => {
     try {
       return typeof content === 'string' ? JSON.parse(content) : content
@@ -79,6 +101,7 @@ export function FileBlock({ content, id, onUpdate, isEditing, metadata }: FileBl
         type="file"
         id={`file-upload-${id}`}
         className="hidden"
+        onChange={handleFileUpload}
       />
       <div className="flex flex-col items-center gap-2">
         <Button
@@ -148,6 +171,7 @@ export function FileBlock({ content, id, onUpdate, isEditing, metadata }: FileBl
               type="file"
               id={`file-change-${id}`}
               className="hidden"
+              onChange={handleFileUpload}
             />
             <Button
               variant="default"
